--- a/lib/jxl/enc_ans.cc
+++ b/lib/jxl/enc_ans.cc
@@ -1870,7 +1870,6 @@ HistogramParams HistogramParams::ForModular(
   }
   if (cparams.decoding_speed_tier >= 2) {
     params.max_histograms = 12;
-  // This barely increases decoding speed, is it even worth using?
   }
   if (cparams.decoding_speed_tier >= 3) {
     params.lz77_method = cparams.speed_tier >= SpeedTier::kCheetah
@@ -1878,14 +1877,6 @@ HistogramParams HistogramParams::ForModular(
                          : cparams.speed_tier >= SpeedTier::kKitten
                              ? HistogramParams::LZ77Method::kLZ77
                              : HistogramParams::LZ77Method::kOptimal;
-  }
-  if (cparams.decoding_speed_tier >= 4) {
-    params.clustering = HistogramParams::ClusteringType::kBest;
-	params.ans_histogram_strategy = HistogramParams::ANSHistogramStrategy::kPrecise;
-  //  This change needs more benchmarking, better density and slightly better
-  //  decoding speed on my PC. Old params are below.
-  //    params.uint_method = HistogramParams::HybridUintMethod::k000;
-  //    params.force_huffman = true;
   }
   return params;
 }