--- a/lib/jxl/enc_frame.cc
+++ b/lib/jxl/enc_frame.cc
@@ -341,6 +341,12 @@ Status MakeFrameHeader(size_t xsize, size_t ysize,
     } else {
       frame_header->group_size_shift = cparams.modular_group_size_shift;
     }
+    if (cparams.modular_group_size_shift == -1 && cparams.decoding_speed_tier >= 2) {
+	  frame_header->group_size_shift = 0;
+	  // by default uses the smallest group size for faster decoding 2 and
+	  // higher, greatly speeds up decoding via multithreading at the cost
+	  // of density.
+	 } else {}
   }
 
   if (jpeg_data) {