--- a/lib/jxl/enc_modular.cc
+++ b/lib/jxl/enc_modular.cc
@@ -380,7 +380,7 @@ void try_palettes(Image& gi, int& max_bitdepth, int& maxval,
       maybe_palette.lossy_palette =
           (cparams_.lossy_palette && maybe_palette.num_c == 3);
       if (maybe_palette.lossy_palette) {
-        maybe_palette.predictor = Predictor::Average4;
+        maybe_palette.predictor = delta_pred_;
       }
       // TODO(veluca): use a custom weighted header if using the weighted
       // predictor.
@@ -401,7 +401,7 @@ void try_palettes(Image& gi, int& max_bitdepth, int& maxval,
       maybe_palette_3.ordered_palette = cparams_.palette_colors >= 0;
       maybe_palette_3.lossy_palette = cparams_.lossy_palette;
       if (maybe_palette_3.lossy_palette) {
-        maybe_palette_3.predictor = Predictor::Average4;
+        maybe_palette_3.predictor = delta_pred_;
       }
       if (maybe_do_transform(gi, maybe_palette_3, cparams_, weighted::Header(),
                              cost_before, pool, cparams_.options.zero_tokens)) {
@@ -604,8 +604,8 @@ Status ModularFrameEncoder::Init(const FrameHeader& frame_header,
       // multipliers in lossy mode.
       cparams_.options.predictor = Predictor::Variable;
     } else if (cparams_.lossy_palette) {
-      // AvgAll predictor for delta palette/lossy palette
-      cparams_.options.predictor = Predictor::Average4;
+      // zero predictor for lossy palette indices
+      cparams_.options.predictor = Predictor::Zero;
     } else if (!cparams_.IsLossless()) {
       // If not responsive and lossy. TODO(veluca): use near_lossless instead?
       cparams_.options.predictor = Predictor::Gradient;
@@ -619,7 +619,10 @@ Status ModularFrameEncoder::Init(const FrameHeader& frame_header,
       // just gradient predictor in thunder mode
       cparams_.options.predictor = Predictor::Gradient;
     }
-  }
+   } else {
+      delta_pred_ = cparams_.options.predictor;
+      if (cparams_.lossy_palette) cparams_.options.predictor = Predictor::Zero;
+   }
   if (!cparams_.ModularPartIsLossless()) {
     if (cparams_.options.predictor == Predictor::Weighted ||
         cparams_.options.predictor == Predictor::Variable ||