### Description

This PR changes the encoding parameters for Lossless when `--faster_decoding` is specified, resulting in up to 80% smaller files and 25% faster decoding.

Progressive Lossless is also improved. Up to 40% smaller files and 20% faster decoding, with an additional 800% encode speed increase, thanks to enabling local MA trees.

These results change depending on encoding parameters, image content, CPU and decoder.
Only libjxl and jxl-oxide were tested, due to jxlatte being too unstable. Oxide proved to be considerably faster for Progressive Lossless and our new level 4 Faster Decoding, suggesting libjxl could be improved by examining it.

Tested on a Ryzen 5600H with a 10 MP image at effort 7.
BPP is now much more inline with the base effort 7 encode, following a logical progression of less density as you increase level.
| BPP | Main | PR |  Δ (%) |
|------|-------|----|--------|
| Default Lossless | 3.458  | 3.457  | N/A |
| Faster Decoding 1 | 15.760 | 3.497 | -77.81 |
| Faster Decoding 2 | 15.760 | 3.578 | -77.30 |
| Faster Decoding 3 | 5.755 | 4.172 | -27.51 |
| Faster Decoding 4 | 5.755  | 4.478 | -22.19 |
| Progressive | 7.948 | 5.047 | -36.50 |

Levels 1-4 now make sense for decoding speed as well, with 3 and 4 now faster than effort 1. Progressive also got a boost.
| Decode (MP/s) | Main | PR |  Δ (%) |
|------|-------|----|--------|
| Default Lossless | 43.01 | 43.01 | N/A |
| Faster Decoding 1 | 134.56 | 76.60  | -43.07 |
| Faster Decoding 2 | 134.56  | 84.31 | -37.34 |
| Faster Decoding 3 | 112.98 | 142.22 | +5.69 |
| Faster Decoding 4 | 112.98 | 169.22 | +25.76  |
| Progressive | 35.41 | 41.72 | +21.24 |

A bonus major speed up for progressive lossless encoding too, using Local MA trees but still buffering the full image.
| Encode (MP/s) | Main | PR |  Δ (%) |
|------|-------|----|--------|
| Progressive | 0.65 | 4.81 | +740 |

### Pull Request Checklist

- [x] **CLA Signed**: Have you signed the [Contributor License Agreement](https://code.google.com/legal/individual-cla-v1.0.html) (individual or corporate, as appropriate)? Only contributions from signed contributors can be accepted.
- [x] **Authors**: Have you considered adding your name to the [AUTHORS](AUTHORS) file?
- [x] **Code Style**: Have you ensured your code adheres to the project's coding style guidelines? You can use `./ci.sh lint` for automatic code formatting.


Please review the full [contributing guidelines](https://github.com/libjxl/libjxl/blob/main/CONTRIBUTING.md) for more details.