@jonnyawsom3 2025-04-20
Tests are failing due to using Local MA trees instead of Global. Discussing on Discord how to fix it, as progressive lossless works with Local but the old behaviour only allows Global.