@Galaxy4594 2025-04-24
Since #4185 (not #4154, adding those changes without the refactor actually made our changes more effective!) our progressive mode changes don't work anymore, it's back to it's broken state. Shouldn't code refactors leave fundamental behavior unchanged? Progressive size balloons with the "none" predictor `-P 0`, which it shouldn't. I'm at a complete loss as to what changed and I don't have the skill and spare time to fix this issue. Any ideas @Melirius?
```
5828 x 4371 photo, effort 7
main baseline ------ 7.625 bpp
main progressive -- 11.700 bpp
after refactor PR - 11.555 bpp
before refactor PR - 8.195 bpp
```