@Melirius 2025-04-24
The only functional changes were correct cost estimations for special-case histograms, following Encode function, so it should not broke anything.