@jonnyawsom3 2025-04-24
> Wow, could you provide the file and exact parameters for me to find out the cause of problem?

We don't think it's from your PR anymore, but if you want to try and help, run both of these images through a build of this fork (or download from Github Actions) as `cjxl -d 0 -p`

This one uses global trees, as it's 2048 wide, which is the limit before buffering is enabled.
![2048](https://github.com/user-attachments/assets/8a3821c7-1886-4d98-aa3c-94224eb8d233)
This one uses local trees, as it's 2049 wide, but density should be equal or better, not worse.
![2049](https://github.com/user-attachments/assets/5b13e023-1c8b-4f9e-bd39-cbeb16051316)