@jonnyawsom3 2025-04-24
Progressive lossless is now fixed, along with fixing Delta Palette at efforts lower than 10.