@jonnyawsom3 2025-04-24
That also didn't fix it... Not sure why the small dimension check is being ignored, using e10 is still too big so it's not local trees, but we've only touched faster decode and progressive lossless. I can't see why it's so much larger. 10% on a 1080p file, 300% on a test