@Melirius 2025-04-25
Have you checked also influence of this fix on density? https://github.com/libjxl/libjxl/pull/4135