@Melirius 2025-04-27
> Going to hold on #4207 before merging with main again, to hopefully pass all checks.

WASM is failing too, hmm...