@Galaxy4594 2025-05-05
@Melirius As it turns out, #4154 is the culprit of the small regression for `--faster_decoding`. I read your PR and I understand why some files have gone up in size, so how could this be fixed?

| Settings | Pre-PR (BPP) | Post-PR (BPP) | Δ (%) |
|-------|-----------|-----------|---------|
| Effort 7 | 7.550 | 7.572 | +0.291
| Faster decoding 1 | 7.630 | 7.651 | +0.275
| Faster decoding 2 | 7.742 | 7.754 | +0.155
| Faster decoding 3 | 8.125 | 8.376 | **+3.089**
| Progressive | 8.195 | 8.194 | -0.012