@Melirius 2025-05-05
Just increase the number of buckets per property, this is the only sane way.