@eustas 2025-07-10
When setting `cparams_.options.wp_tree_mode = ModularOptions::TreeMode::kNoWP;` if predictor is `Weighted` it should be downgraded. Preparing a PR